Settings: MSIM: Fix preferred apn issue for same operator sims
On a multisim device, if same operator sims are used, same preferred apn
is displayed for both subs. When a preferred apn is set for DDS sub,
though preferred apn is null for non DDS sub, since same db entry is
accessed for display, it is displayed as preferred for non DDS sub.

Fix is to unselect apn if it doesnt match preferred apn criteria in
ApnSettings

Change-Id: I5c98a93efa9f2254f1a77d4fb9440338b5efbe64
CRs-Fixed: 2246456
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>