Settings: Fix the preferred APN is not selected issue.
On a multi-SIM device, both SIM cards are camped on 4G. Enter Network settings->Access Point Name,
the default APN is selected. Back to Network settings interface and enter Access Point Name again,
the default APN is not selected.

mSelectedKey is set to null when ApnPreference.unsetChecked is called. So the preferred APN is not
selected when re-enter Access Point Name UI.

The fix is to remove unsetChecked code and set ApnPreference.mSelectedKey static variable according
to preferred APN id per sub.

Change-Id: I25180fc312d9573ce6723af81171722c665e0ada
CRs-Fixed: 2362430
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>